// generated/models/MemeClaimUpdateRequest.ts
// @ts-nocheck
/**
 * 6529.io API
 * This is the API interface description. Brief terminology overview and an authentication example can be found at <a href=\"https://6529.io/about/api\">https://6529.io/about/api</a>.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { MemeClaimAttribute } from '../models/MemeClaimAttribute';
import { HttpFile } from '../http/http';

export class MemeClaimUpdateRequest {
    'season'?: number;
    'image_location'?: string | null;
    'animation_location'?: string | null;
    'metadata_location'?: string | null;
    'edition_size'?: number | null;
    'description'?: string;
    'name'?: string;
    'image_url'?: string | null;
    'attributes'?: Array<MemeClaimAttribute>;
    'animation_url'?: string | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "season",
            "baseName": "season",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "image_location",
            "baseName": "image_location",
            "type": "string",
            "format": ""
        },
        {
            "name": "animation_location",
            "baseName": "animation_location",
            "type": "string",
            "format": ""
        },
        {
            "name": "metadata_location",
            "baseName": "metadata_location",
            "type": "string",
            "format": ""
        },
        {
            "name": "edition_size",
            "baseName": "edition_size",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "image_url",
            "baseName": "image_url",
            "type": "string",
            "format": ""
        },
        {
            "name": "attributes",
            "baseName": "attributes",
            "type": "Array<MemeClaimAttribute>",
            "format": ""
        },
        {
            "name": "animation_url",
            "baseName": "animation_url",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return MemeClaimUpdateRequest.attributeTypeMap;
    }

    public constructor() {
    }
}